requirements.txt
Pygments==2.19.2

PySocks==1.7.1

pyTelegramBotAPI==4.29.1

python-dateutil==2.9.0.post0

python-dotenv==1.1.1

python-telegram-bot==22.5

pytz==2025.2

regex==2025.10.23

requests==2.32.5

rich==14.2.0

schedule==1.2.2

selenium==4.38.0

six==1.17.0

sniffio==1.3.1

sortedcontainers==2.4.0

soupsieve==2.8

telebot==0.0.5

telegram==0.0.1

telegramApi==1.5

tensorboard==2.20.0

tensorboard-data-server==0.7.2

tensorflow==2.20.0

termcolor==3.2.0

tf_keras==2.20.1

tqdm==4.67.1

trio==0.32.0

trio-websocket==0.12.2

typing_extensions==4.15.0

tzdata==2025.2

urllib3==2.5.0



